Added basic pistol FSM that fires and subtracts ammo but still cant find raycasts, added shooting controller to PlayerCamera, added basic tagged enemy "Badman"
Pistol FSM shoots and subtracts ammo and has slots prepped for adding sound and FX but still can't find Raycast targets. Tried using every kind of MousePick, ScreenPick would only be useful if we were doing a standard FPS that casts from the center of the camera and not from the mouse. The Raycast is currently ignoring the UI and Ignore Raycast layers, but it still doesnt seem to capable of picking anything up. The Enemy prefab is just a red box that is tagged as an enemy for the purposes of whenever we can get it seen by the Raycast, and the way its setup currently if the Raycast does see the enemy then it should just destroy the gameobject wholesale.